requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-filter==25.2
psycopg2-binary==2.9.10
python-decouple==3.8
celery==5.3.4
redis==5.0.1
requests==2.32.3
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
